refactor(repository): split android_stage/video_source out of mod.rs (#590)
repository/mod.rs was at 859 prod lines, heading into the 1000-line hard
fail (#586/#587 must edit it next). Extract the android_stage_display and
video_source CRUD groups into sibling files, same pattern already used by
resolume.rs/audit.rs/library.rs: a self-contained impl Repository block per
file, calling the existing pub(super) record_settings_audit_on helper.

Also extract insert_bible_passages_chunked out of
replace_bible_translation_passages (116/120 lines) to get headroom before
#586/#587 land.

Pure code movement, no behavior change. mod.rs drops from 859 to 434 lines;
android_stage.rs 150, video_source.rs 305 — all well under the 800 warn cap.
Existing test suite is the safety net (no test edits).

Closes #590
